Created individual modeles for all shared functions that didn't have a module
