+++ src/lib/orionld/troe/pgGeoLineStringPush.cpp
@@ -36,68 +36,8 @@ extern "C"
 #include "logMsg/traceLevels.h"                                // Lmt*
 
 #include "orionld/common/orionldState.h"                       // orionldState
+#include "orionld/troe/kjGeoLineStringExtract.h"               // kjGeoLineStringExtract
 #include "orionld/troe/pgGeoLineStringPush.h"                  // Own interface
-
-
-
-extern bool kjGeoPointExtract(KjNode* coordinatesP, double* longitudeP, double* latitudeP, double* altitudeP);
-// -----------------------------------------------------------------------------
-//
-// kjGeoLineStringExtract -
-//
-// A LINESTRING looks like this:
-//
-// "value": {
-//   "type": "LineString",
-//   "coordinates": [ [0,0], [4,0], [4,4], [0,4] ]
-// }
-//
-// In PostGIS, the LINESTRING would look like this:
-//   LINESTRING(0 0,  4 0,  4 4,  0 4,  0 0)
-//
-bool kjGeoLineStringExtract(KjNode* coordinatesP, char* lineStringCoordsString, int lineStringCoordsLen)
-{
-  int     lineStringCoordsIx = 0;
-  KjNode* pointP = coordinatesP->value.firstChildP;
-
-
-  while (pointP != NULL)  // Second Array
-  {
-    double  longitude;
-    double  latitude;
-    double  altitude;
-    char    pointBuffer[64];
-
-    if (kjGeoPointExtract(pointP, &longitude, &latitude, &altitude) == false)
-      LM_RE(false, ("Internal Error (unable to extract longitude/latitude/altitude for a Point) "));
-
-    if (altitude != 0)
-      snprintf(pointBuffer, sizeof(pointBuffer), "%f %f %f", longitude, latitude, altitude);
-    else
-      snprintf(pointBuffer, sizeof(pointBuffer), "%f %f", longitude, latitude);
-
-    int pointBufferLen = strlen(pointBuffer);
-
-    if (lineStringCoordsIx + pointBufferLen + 1 >= lineStringCoordsLen)
-      LM_RE(false, ("Not enough room in lineStringCoordsString - fix and recompile"));
-
-    if (lineStringCoordsIx != 0)  // Add a comma before the Point, unless it's the first point
-    {
-      lineStringCoordsString[lineStringCoordsIx] = ',';
-      ++lineStringCoordsIx;
-    }
-
-    LM_TMP(("Appending '%s' to lineStringCoordsString", pointBuffer));
-    strncpy(&lineStringCoordsString[lineStringCoordsIx], pointBuffer, lineStringCoordsLen - lineStringCoordsIx);
-    lineStringCoordsIx += pointBufferLen;
-
-    pointP = pointP->next;
-  }
-
-  LM_TMP(("FINAL lineStringCoordsString: '%s'", lineStringCoordsString));
-
-  return true;
-}
