+++ src/lib/orionld/troe/pgGeoPolygonPush.cpp
@@ -36,101 +36,8 @@ extern "C"
 #include "logMsg/traceLevels.h"                                // Lmt*
 
 #include "orionld/common/orionldState.h"                       // orionldState
+#include "orionld/troe/kjGeoPolygonExtract.h"                  // kjGeoPolygonExtract
 #include "orionld/troe/pgGeoPolygonPush.h"                     // Own interface
-
-
-
-extern bool kjGeoPointExtract(KjNode* coordinatesP, double* longitudeP, double* latitudeP, double* altitudeP);
-// -----------------------------------------------------------------------------
-//
-// kjGeoPolygonExtract -
-//
-// A POLYGON looks like this:
-//
-// "value": {
-//   "type": "Polygon",
-//   "coordinates": [[ [0,0], [4,0], [4,4], [0,4], [0,0] ], [ [1,1], [2,1], [2,2], [1, 2], [1,1] ]]
-// }
-//
-// In PostGIS, the POLYGON would look like this:
-//   POLYGON((0 0,  4 0,  4 4,  0 4,  0 0))
-//
-// A DONUT would look like this:
-//   POLYGON((0 0,  4 0,  4 4,  0 4,  0 0),(1 1,  2 1,  2 2,  1 2,  1 1))
-//
-bool kjGeoPolygonExtract(KjNode* coordinatesP, char* polygonCoordsString, int polygonCoordsLen)
-{
-  KjNode* subPolygonP     = coordinatesP->value.firstChildP;
-  int     polygonCoordsIx = 1;
-
-  // coordinatesP is the toplevel array
-
-  // <DEBUG>
-  char buf[1024];
-  kjRender(orionldState.kjsonP, coordinatesP, buf, sizeof(buf));
-  LM_TMP(("Coordinates: %s", buf));
-  // </DEBUG>
-
-  polygonCoordsString[0] = '(';
-
-  while (subPolygonP != NULL)  // First array
-  {
-    KjNode* pointP = subPolygonP->value.firstChildP;
-
-    while (pointP != NULL)  // Second Array
-    {
-      double  longitude;
-      double  latitude;
-      double  altitude;
-      char    pointBuffer[64];
-
-      // <DEBUG>
-      char buf[1024];
-      kjRender(orionldState.kjsonP, coordinatesP, buf, sizeof(buf));
-      LM_TMP(("Coordinates: %s", buf));
-      kjRender(orionldState.kjsonP, pointP, buf, sizeof(buf));
-      LM_TMP(("Point: %s", buf));
-      // </DEBUG>
-
-      if (kjGeoPointExtract(pointP, &longitude, &latitude, &altitude) == false)
-        LM_RE(false, ("Internal Error (unable to extract longitude/latitude/altitude for a Point) "));
-
-      if (altitude != 0)
-        snprintf(pointBuffer, sizeof(pointBuffer), "%f %f %f", longitude, latitude, altitude);
-      else
-        snprintf(pointBuffer, sizeof(pointBuffer), "%f %f", longitude, latitude);
-
-      int pointBufferLen = strlen(pointBuffer);
-
-      LM_TMP(("polygonCoordsIx  == %d", polygonCoordsIx));
-      LM_TMP(("pointBufferLen   == %d", pointBufferLen));
-      LM_TMP(("polygonCoordsLen == %d", polygonCoordsLen));
-      LM_TMP(("polygonCoordsString: '%s'", polygonCoordsString));
-      if (polygonCoordsIx + pointBufferLen + 1 >= polygonCoordsLen)
-        LM_RE(false, ("Not enough room in polygonCoordsString - fix and recompile"));
-
-      if (polygonCoordsIx != 1)  // Add a comma before the (Point), unless it's the first point
-      {
-        polygonCoordsString[polygonCoordsIx] = ',';
-        ++polygonCoordsIx;
-      }
-
-      LM_TMP(("Appending '%s' to polygonCoordsString", pointBuffer));
-      strncpy(&polygonCoordsString[polygonCoordsIx], pointBuffer, polygonCoordsLen - polygonCoordsIx);
-      polygonCoordsIx += pointBufferLen;
-
-      pointP = pointP->next;
-    }
-
-    subPolygonP = subPolygonP->next;
-  }
-
-  polygonCoordsString[polygonCoordsIx] = ')';
-
-  LM_TMP(("FINAL polygonCoordsString: '%s'", polygonCoordsString));
-
-  return true;
-}
